+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 * This project follows the guidelines outlined on [keepachangelog.com](http://keepachangelog.com/).
 
 ## [Unreleased]
+
+## [0.17.0] - 2016-04-11
 ## Added
 - API method get_filesystem_utilization
 - Automatically-generated uploaded parameter to gcode files